feat: refactors reducers into new file + cleans updater
